reorganized misc components to "OmouComponents" dir, created new background paper component, lessons learned - don't messed with the paper component's root styling
