+++ src/components/FeatureViews/Accounts/TabComponents/ParentContact.js
@@ -1,7 +1,7 @@
 import "./TabComponents.scss";
 import * as hooks from "actions/hooks";
 import Grid from "@material-ui/core/Grid";
-import Loading from "components/Loading";
+import Loading from "components/OmouComponents/Loading";
 import ProfileCard from "../ProfileCard";
 import PropTypes from "prop-types";
 import React from "react";
