+++ src/components/FeatureViews/AdminPortal/UnpaidSessions.js
@@ -7,7 +7,7 @@ import HappyIcon from "@material-ui/icons/SentimentVerySatisfied";
 import Typography from "@material-ui/core/Typography";
 
 import * as hooks from "actions/hooks";
-import Loading from "components/Loading";
+import Loading from "components/OmouComponents/Loading";
 import UnpaidSessionCard from "./UnpaidSessionCard";
 
 const UnpaidSessions = () => {
