+++ src/components/FeatureViews/Registration/registration.scss
@@ -531,7 +531,6 @@
 // Registration Landing
 .RegistrationLanding {
   min-height: 75vh;
-  padding: 2%;
 
   .catalog-setting-wrapper{
     .catalog-setting{
