+++ src/components/FeatureViews/Scheduler/DisplaySessionView.js
@@ -1,19 +1,17 @@
-import React, { useEffect, useMemo, useState } from "react";
+import React, {useEffect, useMemo, useState} from "react";
 // Material UI Imports
 import Grid from "@material-ui/core/Grid";
-import { Link } from "react-router-dom";
+import {NavLink, Redirect, useParams} from "react-router-dom";
 
-import { bindActionCreators } from "redux";
+import {bindActionCreators} from "redux";
 import * as registrationActions from "../../../actions/registrationActions";
-import { useDispatch, useSelector } from "react-redux";
-import { Tooltip, Typography, withStyles } from "@material-ui/core";
-import { NavLink, useParams } from "react-router-dom";
+import {useDispatch, useSelector} from "react-redux";
+import {Tooltip, Typography, withStyles} from "@material-ui/core";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
-import { Redirect } from "react-router-dom"
 import Button from "@material-ui/core/Button";
-import Loading from "../../Loading";
+import Loading from "../../OmouComponents/Loading";
 import Avatar from "@material-ui/core/Avatar";
-import { stringToColor } from "../Accounts/accountUtils";
+import {stringToColor} from "../Accounts/accountUtils";
 import DialogTitle from "@material-ui/core/DialogTitle";
 import Divider from "@material-ui/core/Divider";
 import DialogContent from "@material-ui/core/DialogContent";
@@ -21,22 +19,21 @@ import RadioGroup from "@material-ui/core/RadioGroup";
 import Radio from "@material-ui/core/Radio";
 import DialogActions from "@material-ui/core/DialogActions";
 import Dialog from "@material-ui/core/Dialog";
-import { dayOfWeek } from "../../Form/FormUtils";
+import {dayOfWeek} from "../../Form/FormUtils";
 import * as hooks from "actions/hooks";
 import ConfirmIcon from "@material-ui/icons/CheckCircle";
 import UnconfirmIcon from "@material-ui/icons/Cancel";
-import { EDIT_ALL_SESSIONS, EDIT_CURRENT_SESSION } from "./SessionView";
+import {EDIT_ALL_SESSIONS, EDIT_CURRENT_SESSION} from "./SessionView";
 import DialogContentText from "@material-ui/core/es/DialogContentText";
 import LoadingError from "../Accounts/TabComponents/LoadingCourseError"
-import FormControl from "@material-ui/core/FormControl";
 import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown';
 
 import InstructorSchedule from "../Accounts/TabComponents/InstructorSchedule";
-import SessionPaymentStatusChip from "../../SessionPaymentStatusChip";
-import AddSessions from "components/AddSessions";
+import SessionPaymentStatusChip from "../../OmouComponents/SessionPaymentStatusChip";
+import AddSessions from "components/OmouComponents/AddSessions";
 import Menu from "@material-ui/core/Menu";
 import MenuItem from "@material-ui/core/MenuItem";
-import { capitalizeString } from "../../../utils";
+import {capitalizeString} from "../../../utils";
 
 const StyledMenu = withStyles({
     "paper": {
