+++ src/components/FeatureViews/Scheduler/Scheduler.js
@@ -16,7 +16,6 @@ import FormControl from "@material-ui/core/FormControl";
 import Grid from "@material-ui/core/Grid";
 import IconButton from "@material-ui/core/IconButton";
 import MenuItem from "@material-ui/core/MenuItem";
-import Paper from "@material-ui/core/Paper";
 import Select from "@material-ui/core/Select";
 import TodayIcon from "@material-ui/icons/Today";
 import Tooltip from "@material-ui/core/Tooltip";
@@ -33,7 +32,8 @@ import SessionFilters from "./SessionFilters";
 import {stringToColor} from "../Accounts/accountUtils";
 import {uniques} from "utils";
 import {secondaryFontColor} from "../../../theme/muiTheme";
-import BackButton from "../../BackButton";
+import BackButton from "../../OmouComponents/BackButton";
+import BackgroundPaper from "../../OmouComponents/BackgroundPaper";
 
 const useStyles = makeStyles((theme) => ({
     "bootstrapFormLabel": {
@@ -319,7 +319,7 @@ const Scheduler = () => {
 
     return (
 		<Grid item xs={12} container>
-			<Paper className="scheduler" elevation={2}>
+			<BackgroundPaper className="scheduler" elevation={2}>
 				<Grid item xs={12}>
 					<BackButton/>
 					<hr/>
@@ -486,7 +486,7 @@ const Scheduler = () => {
 									  },
 								  }}/>
 				</Grid>
-			</Paper>
+			</BackgroundPaper>
 		</Grid>
     );
 };
