+++ src/components/OmouComponents/BackgroundPaper.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Paper from "@material-ui/core/Paper";
+
+export default function BackgroundPaper({children, ...rest}) {
+	return (<Paper style={{"padding": "3%"}} {...rest}>
+		{children}
+	</Paper>)
+}
