+++ src/components/OmouComponents/Loading.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import Paper from "@material-ui/core/Paper";
-import Penguin from "./penguin.gif";
+import Penguin from "../penguin.gif";
 import Typography from "@material-ui/core/Typography";
 
 const Loading = ({
